+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- photometric redshift correction now always applied when computing the mean source redshift
+
 ## [0.7.2] - 2023-06-02
 
 ### Changed